widget: add height-for-width hint
Label is the first widget to provide an implementation. Box is the first
container to react to it, as it knows the available width in layout.
Labels now size perfectly when placed in vertical boxes although there
are probably some improvements necessary to calculate the min sizes of
containers (especially how height-for-width is used).