Refactor swipe mechanic
A swipe event has been replaced with a movement event. This event can be
easily transformed from a protected widget method that gives the swipe
direction.

The swipe detector has been replaced with a mouse tracker that tracks
state of the mouse and generates events such that widgets don't have to
maintain most of the state.

There were several other changes that were necessary and at the same
time improved code quality by a lot.